eShop/requirements.txt
django>=2.1.6
django-crispy-forms==1.7.2
gunicorn==19.6.0
html5lib==1.0.1
whitenoise==3.2.1
dj-database-url==0.5.0
psycopg2==2.7.7
pytz==2018.4
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
faker==1.0.5
transliterate==1.10.2
celery==4.1.0
braintree==3.45.0
WeasyPrint==0.42.3
django-tinymce4-lite==1.7.5
django-model-utils==3.1.2
django-environ==0.4.5